+++ src/main/java/com/vn/ivs/ctu/service/ScheduleService.java
@@ -7,9 +7,9 @@ import com.vn.ivs.ctu.entity.Schedule;
 public interface ScheduleService {
 	public long create(Schedule schedule);
 	public List<Schedule> getAll(int idClub);
-	public boolean deleteSchedule(int idSchedule);
+	public boolean deleteSchedule(long idSchedule);
 	public List<Schedule> getAllScheduleTotal();
 	public List<Schedule> getListScheduleAuto(int idClub);
 	public List<Schedule> getListScheduleAuto();
-	public Schedule getScheduleById(int id);
+	public Schedule getScheduleById(long id);
 }
